rename internal ref(TlmCmd -> tlm_cmd): applications
