+++ applications/divided_cmd_utility.c
@@ -12,8 +12,8 @@
  * @note  telemetry_manager などで使っているので，それを参照のこと
  */
 #include "divided_cmd_utility.h"
-#include "../TlmCmd/packet_handler.h"
-#include "../TlmCmd/common_cmd_packet_util.h"
+#include "../tlm_cmd/packet_handler.h"
+#include "../tlm_cmd/common_cmd_packet_util.h"
 #include "../system/time_manager/time_manager.h"
 #include "../system/event_manager/event_logger.h"
 #include "../library/result.h"
