+++ applications/gs_command_dispatcher.h
@@ -5,7 +5,7 @@
 #ifndef GS_COMMAND_DISPATCHER_H_
 #define GS_COMMAND_DISPATCHER_H_
 
-#include "../TlmCmd/command_dispatcher.h"
+#include "../tlm_cmd/command_dispatcher.h"
 #include "../system/application_manager/app_info.h"
 
 extern const CommandDispatcher* const gs_command_dispatcher;
