+++ applications/nop.h
@@ -8,7 +8,7 @@
 #define NOP_H_
 
 #include "../system/application_manager/app_info.h"
-#include "../TlmCmd/common_cmd_packet.h"
+#include "../tlm_cmd/common_cmd_packet.h"
 
 AppInfo NOP_create_app(void);
