+++ applications/realtime_command_dispatcher.c
@@ -1,7 +1,7 @@
 #pragma section REPRO
 #include "realtime_command_dispatcher.h"
-#include "../TlmCmd/packet_handler.h"
-#include "../TlmCmd/common_cmd_packet_util.h"
+#include "../tlm_cmd/packet_handler.h"
+#include "../tlm_cmd/common_cmd_packet_util.h"
 #include "../library/result.h"
 
 static CommandDispatcher realtime_command_dispatcher_;
