+++ applications/telemetry_manager.h
@@ -7,9 +7,9 @@
 #define TELEMETRY_MANAGER_H_
 
 #include "../system/application_manager/app_info.h"
-#include "../TlmCmd/common_cmd_packet.h"
-#include "../TlmCmd/common_cmd_packet_util.h"
-#include "../TlmCmd/block_command_table.h"
+#include "../tlm_cmd/common_cmd_packet.h"
+#include "../tlm_cmd/common_cmd_packet_util.h"
+#include "../tlm_cmd/block_command_table.h"
 #include <src_user/TlmCmd/telemetry_definitions.h>
 #include <src_user/Settings/TlmCmd/Ccsds/apid_define.h>
