+++ applications/test_app/test_ccp_util.h
@@ -5,7 +5,7 @@
 #ifndef TEST_CCP_UTIL_H_
 #define TEST_CCP_UTIL_H_
 
-#include "../../TlmCmd/common_cmd_packet.h"
+#include "../../tlm_cmd/common_cmd_packet.h"
 
 CCP_CmdRet Cmd_TEST_CCP_REGISTER_TLC_ASAP(const CommonCmdPacket* packet);
 CCP_CmdRet Cmd_TEST_CCP_GET_RAW_PARAM_INFO(const CommonCmdPacket* packet);
