+++ applications/timeline_command_dispatcher.c
@@ -1,9 +1,9 @@
 #pragma section REPRO
 #include "timeline_command_dispatcher.h"
-#include "../TlmCmd/packet_handler.h"
+#include "../tlm_cmd/packet_handler.h"
 #include "../system/time_manager/time_manager.h"
 #include "../system/event_manager/event_logger.h"
-#include "../TlmCmd/common_cmd_packet_util.h"
+#include "../tlm_cmd/common_cmd_packet_util.h"
 #include "../library/result.h"
 
 #include <string.h> // for memset
